GSM Progress 7
We now have mortgages, a treasure none of us will probably possess given house prices are rising higher than wages. But we have the closest thing to it here in property tycoon! Note, I haven't actually done code for when a player lands on something, right now it's a pretty chill game of pass go and buy funny squares. - E