Property Development is here!
Real estate sim time guys, we take the code that does the house/hotel "building" and actually get it to build houses and hotels! It actually looks like a property tycoon board now kinda? A little unrelated but it's now light outside which is really gonna fuck with my sleep lol but we move. - E